Share common logic bettween `create` and `import` commands
